+++ packages/picasso.js/src/api/components.js
@@ -14,6 +14,7 @@ import categoricalLegend from '../core/chart-components/legend-cat';
 import sequentialLegend from '../core/chart-components/legend-seq';
 import lineComponent from '../core/chart-components/line';
 import brushAreaComponent from '../web/components/brush-area';
+import tooltipComponent from '../web/components/tooltip';
 
 export default [
   boxComponent,
@@ -31,5 +32,6 @@ export default [
   categoricalLegend,
   sequentialLegend,
   lineComponent,
-  brushAreaComponent
+  brushAreaComponent,
+  tooltipComponent
 ];
